Add clean avro option
Refactor code according to comments
