+++ utils/serde/src/main/java/io/apicurio/registry/utils/serde/AbstractKafkaSerDe.java
@@ -22,6 +22,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.Consumer;
 
+import io.apicurio.registry.rest.beans.ArtifactMetaData;
+import io.apicurio.registry.rest.beans.VersionMetaData;
 import io.apicurio.registry.utils.serde.util.HeaderUtils;
 import org.apache.kafka.common.errors.SerializationException;
 import org.slf4j.Logger;
@@ -52,7 +54,7 @@ public abstract class AbstractKafkaSerDe<T extends AbstractKafkaSerDe<T>> implem
     public static final String REGISTRY_CONFLUENT_ID_HANDLER_CONFIG_PARAM = "apicurio.registry.as-confluent";
 
     // Constants for using headers to store the ids
-    public static final String USE_HEADERS = "apicurio.use.headers";
+    public static final String USE_HEADERS = "apicurio.registry.use.headers";
 
     public static final byte MAGIC_BYTE = 0x0;
     protected boolean key; // do we handle key or value with this ser/de?
@@ -184,5 +186,24 @@ public abstract class AbstractKafkaSerDe<T extends AbstractKafkaSerDe<T>> implem
     public Object setKey(boolean key) {
         this.key = key;
         return self();
+    }
+
+    /**
+     * Converts an artifact id and version to a global id by querying the registry.  If anything goes wrong,
+     * throws an appropriate exception.
+     * @param artifactId
+     * @param version
+     */
+    protected Long toGlobalId(String artifactId, Integer version) {
+        if (artifactId == null) {
+            throw new RuntimeException("ArtifactId not found in headers.");
+        }
+        if (version == null) {
+            ArtifactMetaData amd = getClient().getArtifactMetaData(artifactId);
+            return amd.getGlobalId();
+        } else {
+            VersionMetaData vmd = getClient().getArtifactVersionMetaData(version, artifactId);
+            return vmd.getGlobalId();
+        }
     }
 }
